switched from double representation to integer representation this fixed bugs in collisoin detection

+++ Jump_n_Bump/src/de/jumpnbump/usecases/game/factories/WallFactory.java
@@ -5,14 +5,14 @@ import de.jumpnbump.usecases.game.model.Wall;
 
 public class WallFactory {
 
-	public static Wall createWall(double x, double y) {
+	public static Wall createWall(int x, int y) {
 		int id = IdCounter.getNextId();
 		Wall wall = new Wall(id, x, y, ModelConstants.WALL_WIDTH,
 				ModelConstants.WALL_HEIGHT);
 		return wall;
 	}
 
-	public static Wall createWall(double x, double y, double maxX, double maxY) {
+	public static Wall createWall(int x, int y, int maxX, int maxY) {
 		int id = IdCounter.getNextId();
 		Wall wall = new Wall(id, x, y, maxX, maxY);
 		return wall;
