+++ Jump_n_Bump/src/de/jumpnbump/usecases/game/model/IcyWall.java
@@ -2,7 +2,7 @@ package de.jumpnbump.usecases.game.model;
 
 public class IcyWall extends Wall {
 
-	public IcyWall(int id, double minX, double minY, double maxX, double maxY) {
+	public IcyWall(int id, int minX, int minY, int maxX, int maxY) {
 		super(id, minX, minY, maxX, maxY, 0xff8888ff);
 	}
